cheri: riscv: Use uintptr_t in arch/riscv/kernel/probes/
Use uintptr_t instead of unsigned long in arch/riscv/kernel/probes/.

+++ arch/riscv/include/asm/probes.h
@@ -4,14 +4,14 @@
 #define _ASM_RISCV_PROBES_H
 
 typedef u32 probe_opcode_t;
-typedef bool (probes_handler_t) (u32 opcode, unsigned long addr, struct pt_regs *);
+typedef bool (probes_handler_t) (u32 opcode, uintptr_t addr, struct pt_regs *);
 
 /* architecture specific copy of original instruction */
 struct arch_probe_insn {
 	probe_opcode_t *insn;
 	probes_handler_t *handler;
 	/* restore address after simulation */
-	unsigned long restore;
+	uintptr_t restore;
 };
 
 #ifdef CONFIG_KPROBES
