+++ arch/riscv/kernel/probes/rethook.h
@@ -2,7 +2,7 @@
 #ifndef __RISCV_RETHOOK_H
 #define __RISCV_RETHOOK_H
 
-unsigned long arch_rethook_trampoline_callback(struct pt_regs *regs);
+uintptr_t arch_rethook_trampoline_callback(struct pt_regs *regs);
 void arch_rethook_prepare(struct rethook_node *rhn, struct pt_regs *regs, bool mcount);
 
 #endif
